Compilation & Execution

Compilation & Execution

CUDAnative.@cuda — Macro.

@cuda [kwargs...] func(args...)

High-level interface for executing code on a GPU. The @cuda macro should prefix a call, with func a callable function or object that should return nothing. It will be compiled to a CUDA function upon first use, and to a certain extent arguments will be converted and managed automatically (see cudaconvert). Finally, a call to CUDAdrv.cudacall is performed, scheduling a kernel launch on the current CUDA context.

Several keyword arguments are supported that influence kernel compilation and execution. For more information, refer to the documentation of respectively cufunction and CUDAnative.Kernel

The underlying operations (argument conversion, kernel compilation, kernel call) can be performed explicitly when more control is needed, e.g. to reflect on the resource usage of a kernel to determine the launch configuration:

args = ...
GC.@preserve args begin
    kernel_args = cudaconvert.(args)
    kernel_tt = Tuple{Core.Typeof.(kernel_args)...}
    kernel = CUDAnative.cufunction(f, kernel_tt; compilation_kwargs)
    kernel(kernel_args...; launch_kwargs)
end

source

CUDAnative.cudaconvert — Function.

cudaconvert(x)

Low-level interface to convert values to a representation that is GPU compatible. For a higher-level interface, use @cuda.

By default, CUDAnative does only provide a minimal set of conversions for elementary types such as tuples. If you need your type to convert before execution on a GPU, be sure to add methods to this function.

For the time being, conversions for CUDAdrv.CuArray objects are also provided, returning a corresponding CuDeviceArray object in global memory. This will be deprecated in favor of functionality from the CuArrays.jl package.

source

CUDAnative.nearest_warpsize — Function.

nearest_warpsize(dev::CuDevice, threads::Integer)

Return the nearest number of threads that is a multiple of the warp size of a device.

This is a common requirement, eg. when using shuffle intrinsics.

source